requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
dask==2025.1.0
distributed==2025.1.0
docstring_parser==0.16
filelock==3.17.0
fonttools==4.56.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.2
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
shapely==2.0.7
six==1.17.0
sortedcontainers==2.4.0
sympy==1.13.1
tblib==3.0.0
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vertexai==1.71.1
zict==3.0.0
zipp==3.21.0
